fix: address meroreviewer feedback
- Remove unnecessary #[non_exhaustive] from CrdtType (internal enum)
- Remove wildcard match in merodb/abi.rs (no longer needed)
- Fix PnCounter comment (uses two maps, not one)
- Remove redundant i64 cast in get_pn_counter

+++ apps/e2e-kv-store/src/lib.rs
@@ -882,7 +882,7 @@ impl E2eKvStore {
         self.crdt_pn_counters
             .get(&key)
             .map_err(|e| format!("Get failed: {:?}", e))?
-            .map(|c| c.value().unwrap_or(0) as i64)
+            .map(|c| c.value().unwrap_or(0))
             .ok_or_else(|| "PNCounter not found".to_owned())
     }
